Migrate from Kotlin synthetics to Jetpack view binding
