+++ tasks/android-kotlin/ToDo/app/build.gradle
@@ -35,6 +35,10 @@ android {
         jvmTarget = "1.8"
     }
 
+    buildFeatures {
+        viewBinding true
+    }
+
     composeOptions {
         kotlinCompilerExtensionVersion "1.4.3"
     }
@@ -48,13 +52,12 @@ android {
 
 dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
-    implementation ("org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version")
-    implementation ("androidx.appcompat:appcompat:1.2.0")
-    implementation ("androidx.core:core-ktx:1.3.2")
-    implementation ("androidx.constraintlayout:constraintlayout:2.0.1")
-    implementation ("com.google.android.material:material:1.2.1")
-    testImplementation ("junit:junit:4.12")
-    androidTestImplementation ("androidx.test:runner:1.3.0")
-    androidTestImplementation ("androidx.test.espresso:espresso-core:3.3.0")
-    implementation ("live.ditto:ditto:3.0.0")
+    implementation ("androidx.appcompat:appcompat:1.6.1")
+    implementation ("androidx.core:core-ktx:1.10.0")
+    implementation ("androidx.constraintlayout:constraintlayout:2.1.4")
+    implementation ("com.google.android.material:material:1.8.0")
+    testImplementation ("junit:junit:4.13.2")
+    androidTestImplementation ("androidx.test:runner:1.5.2")
+    androidTestImplementation ("androidx.test.espresso:espresso-core:3.5.1")
+    implementation ("live.ditto:ditto:4.0.1")
 }
