+++ tasks/android-kotlin/ToDo/app/src/main/java/com/dittolive/todo/NewTaskDialogFragment.kt
@@ -6,8 +6,11 @@ import android.app.Dialog
 import android.os.Bundle
 import android.widget.TextView
 import androidx.fragment.app.DialogFragment
+import live.dittolive.todo.R
 
 class NewTaskDialogFragment: DialogFragment() {
+
+
 
     interface NewTaskDialogListener {
         fun onDialogSave(dialog: DialogFragment, task: String)
@@ -17,25 +20,24 @@ class NewTaskDialogFragment: DialogFragment() {
     var newTaskDialogListener: NewTaskDialogListener? = null
 
     companion object {
-        fun newInstance(title: Int): NewTaskDialogFragment {
+        fun newInstance(): NewTaskDialogFragment {
             val newTaskDialogFragment = NewTaskDialogFragment()
             val args = Bundle()
-            args.putInt("dialog_title", title)
             newTaskDialogFragment.arguments = args
             return newTaskDialogFragment
         }
     }
 
     override fun onCreateDialog(savedInstanceState: Bundle?): Dialog { // 5
-        val title = arguments!!.getInt("dialog_title")
+        val title = "Add New Task"
         val builder = AlertDialog.Builder(activity)
         builder.setTitle(title)
 
-        val dialogView = activity!!.layoutInflater.inflate(R.layout.dialog_new_task, null)
-        val task = dialogView.findViewById<TextView>(R.id.editText)
+        val dialogView = activity?.layoutInflater?.inflate(R.layout.dialog_new_task, null)
+        val task = dialogView?.findViewById<TextView>(R.id.editText)
 
         builder.setView(dialogView)
-            .setPositiveButton(R.string.save) { _, _ -> newTaskDialogListener?.onDialogSave(this, task.text.toString()) }
+            .setPositiveButton(R.string.save) { _, _ -> newTaskDialogListener?.onDialogSave(this, task?.text.toString()) }
             .setNegativeButton(android.R.string.cancel) { _, _ -> newTaskDialogListener?.onDialogCancel(this) }
         return builder.create()
     }
