+++ tasks/android-kotlin/ToDo/app/src/main/java/com/dittolive/todo/TaskViewHolder.kt
@@ -1,0 +1,6 @@
+package com.dittolive.todo
+
+import androidx.recyclerview.widget.RecyclerView
+import live.dittolive.todo.databinding.TaskViewBinding
+
+class TaskViewHolder(val binding: TaskViewBinding): RecyclerView.ViewHolder(binding.root)
