+++ tasks/android-kotlin/ToDo/app/src/main/res/values/strings.xml
@@ -4,4 +4,5 @@
     <string name="title_activity_main">ToDo</string>
     <string name="add_new_task_dialog_title">Add New Task</string>
     <string name="save">Save</string>
+    <string name="textview">TextView</string>
 </resources>
